Move these helpers to Core.Networking
